fix the incorrect havoc (should be Integer value not Long)

+++ trunk/source/Library-ModelCheckerAssistant/src/tw/ntu/svvrl/ultimate/lib/modelcheckerassistant/transitiontoolkit/threadtransitiontoolkit/ThreadStatementsExecutor.java
@@ -245,7 +245,7 @@ public class ThreadStatementsExecutor extends StatementsExecutor<ThreadState> {
 						updateThreadState(procName, varName, value);
 						break;
 					case BoogiePrimitiveType.INT:
-						value = r.nextLong();
+						value = Long.valueOf(r.nextInt());
 						updateThreadState(procName, varName, value);
 						break;
 					case BoogiePrimitiveType.REAL:
